scraper/requirements.txt
httpx==0.25.2
parsel==1.8.1
supabase==2.4.0
python-dotenv==1.0.0
python-dateutil==2.8.2
pydantic==2.5.0